src/requirements.shared.txt
boto3>=1.34.0
openai>=1.0.0
google-api-python-client>=2.110.0
python-dotenv>=1.0.0
pytest==8.4.1